[intel-npu][manual rebase] refactor options and config
