+++ src/plugins/intel_npu/src/al/include/intel_npu/config/npuw.hpp
@@ -6,7 +6,7 @@
 
 #include <thread>
 
-#include "common.hpp"
+#include "intel_npu/config/options.hpp"
 #include "intel_npu/npu_private_properties.hpp"
 #include "intel_npu/npuw_private_properties.hpp"
 
@@ -19,13 +19,19 @@ namespace intel_npu {
 void registerNPUWOptions(OptionsDesc& desc);
 void registerNPUWLLMOptions(OptionsDesc& desc);
 
-#define DEFINE_OPT(Name, Type, DefaultValue, PropertyKey, Mode)                     \
-    struct Name final : OptionBase<Name, Type> {                                    \
-        static std::string_view key() { return ov::intel_npu::PropertyKey.name(); } \
-                                                                                    \
-        static Type defaultValue() { return DefaultValue; }                         \
-                                                                                    \
-        static OptionMode mode() { return OptionMode::Mode; }                       \
+#define DEFINE_OPT(Name, Type, DefaultValue, PropertyKey, Mode) \
+    struct Name final : OptionBase<Name, Type> {                \
+        static std::string_view key() {                         \
+            return ov::intel_npu::PropertyKey.name();           \
+        }                                                       \
+                                                                \
+        static Type defaultValue() {                            \
+            return DefaultValue;                                \
+        }                                                       \
+                                                                \
+        static OptionMode mode() {                              \
+            return OptionMode::Mode;                            \
+        }                                                       \
     };
 
 DEFINE_OPT(NPU_USE_NPUW, bool, false, use_npuw, CompileTime);
